+++ .gitignore
@@ -34,9 +34,11 @@ node_modules
 
 # Transmute Framework
 
+build
 config.js
 service-account.json
 scratch
 
 sigmate-v3-ti.json
 src/creds.js
+
